Move stateful IoSourceState to seperate file
So that it can be used by the event_poll implementation.
